requirements.txt
django==5.1.1
pillow==10.4.0
moviepy==1.0.3
numpy==2.1.1